优化代码逻辑
