+++ src/capabilities/tools/media/generation.js
@@ -1,0 +1,60 @@
+import fs from 'fs'
+import path from 'path'
+import { nowTimestamp } from '../../../time.js'
+import { emitEvent } from '../../../events.js'
+import { callCapability } from '../../../providers/registry.js'
+import { isDailyLimitReached } from '../../../quota.js'
+import { SANDBOX_ROOT } from '../../sandbox.js'
+
+// generate_lyrics：生成歌词
+export async function execGenerateLyrics({ prompt, mode }) {
+  if (!prompt) return '错误：未提供创作方向'
+  if (isDailyLimitReached('lyrics')) return '错误：今日歌词生成配额已用完'
+
+  const result = await callCapability('lyrics', { prompt, mode })
+
+  // 自动保存歌词到 sandbox
+  const ts = nowTimestamp().replace(/[:.+]/g, '-').slice(0, 19)
+  const fname = `lyrics_${ts}.txt`
+  const content = `# ${result.title}\n风格：${result.style}\n\n${result.lyrics}`
+  const resolved = path.resolve(SANDBOX_ROOT, 'lyrics', fname)
+  fs.mkdirSync(path.dirname(resolved), { recursive: true })
+  fs.writeFileSync(resolved, content, 'utf-8')
+
+  emitEvent('lyrics_created', { path: `lyrics/${fname}`, title: result.title })
+  return `歌词已生成并保存至 lyrics/${fname}\n\n标题：${result.title}\n风格：${result.style}\n\n${result.lyrics}`
+}
+
+// generate_music：生成音乐
+export async function execGenerateMusic({ prompt, lyrics, instrumental }) {
+  if (!prompt) return '错误：未提供音乐描述'
+  if (isDailyLimitReached('music')) return '错误：今日音乐生成配额已用完'
+
+  const result = await callCapability('music', { prompt, lyrics, instrumental })
+
+  const ts = nowTimestamp().replace(/[:.+]/g, '-').slice(0, 19)
+  const fname = `music_${ts}.mp3`
+  const resolved = path.resolve(SANDBOX_ROOT, 'music', fname)
+  fs.mkdirSync(path.dirname(resolved), { recursive: true })
+  fs.writeFileSync(resolved, result.buffer)
+
+  const relPath = `music/${fname}`
+  emitEvent('music_created', { path: relPath, prompt: prompt.slice(0, 60) })
+  console.log(`[music] 已生成: ${relPath}`)
+  return `音乐已生成：${relPath}（时长约 ${result.duration ?? '?'} 秒）`
+}
+
+// generate_image：生成图片
+export async function execGenerateImage({ prompt, aspect_ratio = '1:1', n = 1 }) {
+  if (!prompt) return '错误：未提供图片描述'
+  if (isDailyLimitReached('image')) return '错误：今日图片生成配额已用完（50 次/天）'
+  const validRatios = new Set(['1:1', '16:9', '4:3', '3:4', '9:16'])
+  const ratio = validRatios.has(aspect_ratio) ? aspect_ratio : '1:1'
+  const count = Math.min(Math.max(Math.floor(n) || 1, 1), 4)
+
+  const result = await callCapability('image', { prompt, aspect_ratio: ratio, n: count })
+
+  emitEvent('image_created', { urls: result.urls, prompt: prompt.slice(0, 60) })
+  console.log(`[image] 已生成 ${result.urls.length} 张图片`)
+  return `图片已生成（${result.urls.length} 张）：\n${result.urls.join('\n')}`
+}
